added testing for react components
