+++ .gitignore
@@ -1,1 +1,4 @@
 node_modules/
+/test/__snapshots__
+.babelrc
+enzyme.config.js
